<REFACTOR> 불필요한 user.repository 패키에서 MemoryAppUserRepository 상위 패키지로 옮김
